requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
--extra-index-url https://download.pytorch.org/whl/cu126

aiofiles==23.2.1
    # via gradio
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   gradio
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.5
    # via requests
click==8.3.1
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
faiss-cpu==1.13.2
    # via -r requirements.in
fastapi==0.112.2
    # via
    #   -r requirements.in
    #   gradio
ffmpy==1.0.0
    # via gradio
filelock==3.25.2
    # via
    #   huggingface-hub
    #   torch
fsspec==2026.2.0
    # via
    #   gradio-client
    #   huggingface-hub
    #   torch
gradio==5.1.0
    # via -r requirements.in
gradio-client==1.4.0
    # via gradio
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.4.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.27.2
    # via
    #   -r requirements.in
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langgraph-sdk
    #   langsmith
huggingface-hub==1.7.1
    # via
    #   gradio
    #   gradio-client
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   gradio
    #   torch
joblib==1.5.3
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==1.2.19
    # via
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langgraph==1.1.2
    # via -r requirements.in
langgraph-checkpoint==4.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.8
    # via langgraph
langgraph-sdk==0.3.11
    # via langgraph
langsmith==0.7.17
    # via langchain-core
markdown-it-py==4.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via
    #   -r requirements.in
    #   torch
numpy==2.1.1
    # via
    #   -r requirements.in
    #   faiss-cpu
    #   gradio
    #   pandas
    #   rank-bm25
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
orjson==3.11.7
    # via
    #   -r requirements.in
    #   gradio
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   faiss-cpu
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   transformers
pandas==2.3.3
    # via gradio
pillow==10.4.0
    # via gradio
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
    #   gradio
    #   langchain-core
    #   langgraph
    #   langsmith
    #   pydantic-settings
pydantic-core==2.20.1
    # via pydantic
pydantic-settings==2.4.0
    # via -r requirements.in
pydub==0.25.1
    # via gradio
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.22
    # via gradio
pytz==2026.1.post1
    # via pandas
pyyaml==6.0.3
    # via
    #   gradio
    #   huggingface-hub
    #   langchain-core
    #   transformers
    #   uvicorn
rank-bm25==0.2.2
    # via -r requirements.in
regex==2026.2.28
    # via transformers
requests==2.32.5
    # via
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rich==14.3.3
    # via typer
ruff==0.15.6
    # via gradio
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.17.1
    # via
    #   scikit-learn
    #   sentence-transformers
semantic-version==2.10.0
    # via gradio
sentence-transformers==5.3.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via httpx
sqlalchemy==2.0.48
    # via -r requirements.in
starlette==0.38.6
    # via fastapi
sympy==1.14.0
    # via torch
tenacity==9.1.4
    # via langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
tomlkit==0.12.0
    # via gradio
torch==2.10.0+cu126
    # via sentence-transformers
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==5.3.0
    # via
    #   -r requirements.in
    #   sentence-transformers
typer==0.24.1
    # via
    #   gradio
    #   huggingface-hub
    #   transformers
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   sqlalchemy
    #   torch
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
uvicorn[standard]==0.30.6
    # via
    #   -r requirements.in
    #   gradio
watchfiles==1.1.1
    # via uvicorn
websockets==12.0
    # via
    #   gradio-client
    #   uvicorn
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
zstandard==0.25.0
    # via langsmith
